feat: Correct the MPC model to Global Kinematic Model
Add params class, to change parameters quickly.
Convert velocity to m/s from MPH.
Convert the sign of the steering angle obtained from simulator.
Predict the location of the vehicle at 100ms in the future.
Add costs for steering angle and throttle.
Add Cost of the value between the cte and the slope of the polynomial to start turns in advance.